[level 0] Title: 두 수의 나눗셈, Time: 0.04 ms, Memory: 81 MB -BaekjoonHub

+++ 프로그래머스/lv0/120806. 두 수의 나눗셈/두 수의 나눗셈.java
@@ -1,5 +1,5 @@
 class Solution {
-    public double solution(int num1, int num2) {
+    public int solution(int num1, int num2) {
         double answer = 0;
         answer = (double)num1 / num2 * 1000;
